Final_project/requirements.txt
streamlit
spacy
sentence-transformers
pdfplumber
python-docx
pandas
plotly
